+++ src/requests/getTennant.js
@@ -1,9 +1,10 @@
 import axios from "axios";
 
 const getTennant = async () => {
-  const path = `https://ding-dong-backend.herokuapp.com/tennants/`;
+  const cors = require("cors");
+  const path = `https://ding-dong-backend.herokuapp.com/tennants`;
   try {
-    const response = await axios.get(path);
+    const response = await axios.get(path, cors());
     return response;
   } catch (error) {
     console.log(error.response);
